Created a player class
In this commit I created a player class and instanciated two players. The player class is responsible for drawing PLAYER_ONE to the left, and PLAYER_TWO to the right. It also takes user input and moves the players accordingly. It ensures that the player can't go out of bounds.