Implemented a ball class
In this commit I added a ball class that is responsible for drawing the ball and maintaining its logic. The ball class has to ensure that the ball deflects of the walls and that the ball knows when it gets out of bounds as then we know a goal was scored.